Add pair_kinship on graphs and views as the pinned float32 recurrence (slice 5a)
PedigreeGraph.pair_kinship and PedigreeView.pair_kinship take two row
arrays, one RelationshipPairBlock, or a whole RelationshipPairs, and
return read-only float32 from the ADR 0009 recurrence: peel the deeper
endpoint, ties by the greater row, one float32 rounding per half-sum.
Within a receiver the values are bit-identical to the kinship matrix,
to the reversed endpoint order, and to themselves before and after a
matrix is cached; the cached-matrix sampling branch is deleted (issue
#6).  Blocks from another receiver fail with coordinate_space_mismatch,
row arguments with the pair_* codes, and a memo past its cap with
ResourceError("memo_capacity_exceeded").

The kernel keeps running in the private depth-major order, where the
depth-then-row rule is simply the greater row: reading depth per node
in graph space measured 7 to 9 percent slower on random_30k, and int32
inputs another 6 percent, so the kernel takes int64 depth-major arrays.
Interleaved medians of three against the 0.7.1 kernel on random_30k:
self pairs 31.4 s vs 31.9 s, degree-3 pairs 88.5 s vs 91.3 s, with
kernel RSS 289 MB vs 384 MB and 562 MB vs 747 MB.  The pure-Python
oracle stays in graph space with explicit depth so the kernel parity
test also checks the translation.

compute_pair_kinship keeps its dict form and caller-space rows but now
returns float32 without committing the thread budget (0.8.0-DELETE).
The 0.7.1 parity test exempts only deep_inbred_60g pair kinship and
checks it against the cross-order envelope instead; the row-order test
drops its deep-inbred bit-parity skip.

+++ tests/test_inbreeding_kernel.py
@@ -40,7 +40,7 @@ def _F_via_pairwise(m, f, tw):
     f = np.asarray(f, dtype=np.int32)
     tw = np.asarray(tw, dtype=np.int32)
     rows = np.arange(len(m), dtype=np.int64)
-    return 2.0 * pairwise_kinship(m, f, tw, rows, rows) - 1.0
+    return 2.0 * pairwise_kinship(m, f, tw, rows, rows).astype(np.float64) - 1.0
 
 
 def _F_via_matrix(m, f, tw, gen, n=None):
